KPI_Automation/requirements.txt
requests
pytz
pycryptodome